+++ src/aks-preview/azext_aks_preview/_validators.py
@@ -793,3 +793,10 @@ def validate_azure_service_mesh_revision(namespace):
     found = asm_revision_regex.findall(revision)
     if not found:
         raise InvalidArgumentValueError(f"Revision {revision} is not supported by the service mesh add-on.")
+
+
+def validate_artifact_streaming(namespace):
+    """Validates that artifact streaming enablement can only be used on Linux."""
+    if namespace.enable_artifact_streaming:
+        if hasattr(namespace, 'os_type') and str(namespace.os_type).lower() == "windows":
+            raise ArgumentUsageError('--enable-artifact-streaming can only be set for Linux nodepools')