+++ src/aks-preview/azext_aks_preview/tests/latest/test_agentpool_decorator.py
@@ -243,6 +243,41 @@ class AKSPreviewAgentPoolContextCommonTestCase(unittest.TestCase):
         with self.assertRaises(MutuallyExclusiveArgumentError):
             ctx_2.get_disable_custom_ca_trust()
 
+    def common_get_enable_artifact_streaming(self):
+        # default
+        ctx_1 = AKSPreviewAgentPoolContext(
+            self.cmd,
+            AKSAgentPoolParamDict({"enable_artifact_streaming": None}),
+            self.models,
+            DecoratorMode.CREATE,
+            self.agentpool_decorator_mode,
+        )
+        self.assertEqual(ctx_1.get_enable_artifact_streaming(), None)
+        agentpool_1 = self.create_initialized_agentpool_instance(
+            artifact_streaming_profile=self.models.AgentPoolArtifactStreamingProfile(
+                enabled=True
+            )
+        )
+        ctx_1.attach_agentpool(agentpool_1)
+        self.assertEqual(ctx_1.get_enable_artifact_streaming(), True)
+
+        # default
+        ctx_2 = AKSPreviewAgentPoolContext(
+            self.cmd,
+            AKSAgentPoolParamDict({"enable_artifact_streaming": None}),
+            self.models,
+            DecoratorMode.UPDATE,
+            self.agentpool_decorator_mode,
+        )
+        self.assertEqual(ctx_2.get_enable_artifact_streaming(), None)
+        agentpool_2 = self.create_initialized_agentpool_instance(
+            artifact_streaming_profile=self.models.AgentPoolArtifactStreamingProfile(
+                enabled=True
+            )
+        )
+        ctx_2.attach_agentpool(agentpool_2)
+        self.assertEqual(ctx_2.get_enable_artifact_streaming(), None)
+
 
 class AKSPreviewAgentPoolContextStandaloneModeTestCase(AKSPreviewAgentPoolContextCommonTestCase):
     def setUp(self):
@@ -275,6 +310,8 @@ class AKSPreviewAgentPoolContextStandaloneModeTestCase(AKSPreviewAgentPoolContex
     def test_get_disable_custom_ca_trust(self):
         self.common_get_disable_custom_ca_trust()
 
+    def test_get_enable_artifact_streaming(self):
+        self.common_get_enable_artifact_streaming()
 
 class AKSPreviewAgentPoolContextManagedClusterModeTestCase(AKSPreviewAgentPoolContextCommonTestCase):
     def setUp(self):
@@ -307,6 +344,8 @@ class AKSPreviewAgentPoolContextManagedClusterModeTestCase(AKSPreviewAgentPoolCo
     def test_get_disable_custom_ca_trust(self):
         self.common_get_disable_custom_ca_trust()
 
+    def test_get_enable_artifact_streaming(self):
+        self.common_get_enable_artifact_streaming()
 
 class AKSPreviewAgentPoolAddDecoratorCommonTestCase(unittest.TestCase):
     def _remove_defaults_in_agentpool(self, agentpool):
@@ -430,6 +469,27 @@ class AKSPreviewAgentPoolAddDecoratorCommonTestCase(unittest.TestCase):
         )
         self.assertEqual(dec_agentpool_1, ground_truth_agentpool_1)
 
+    def common_set_up_artifact_streaming(self):
+        dec_1 = AKSPreviewAgentPoolAddDecorator(
+            self.cmd,
+            self.client,
+            {"enable_artifact_streaming": True},
+            self.resource_type,
+            self.agentpool_decorator_mode,
+        )
+        # fail on passing the wrong agentpool object
+        with self.assertRaises(CLIInternalError):
+            dec_1.set_up_artifact_streaming(None)
+        agentpool_1 = self.create_initialized_agentpool_instance(restore_defaults=False)
+        dec_1.context.attach_agentpool(agentpool_1)
+        dec_agentpool_1 = dec_1.set_up_artifact_streaming(agentpool_1)
+        dec_agentpool_1 = self._restore_defaults_in_agentpool(dec_agentpool_1)
+        ground_truth_agentpool_1 = self.create_initialized_agentpool_instance(
+            artifact_streaming_profile=self.models.AgentPoolArtifactStreamingProfile(
+                enabled=True
+            )
+        )
+        self.assertEqual(dec_agentpool_1, ground_truth_agentpool_1)
 
 class AKSPreviewAgentPoolAddDecoratorStandaloneModeTestCase(AKSPreviewAgentPoolAddDecoratorCommonTestCase):
     def setUp(self):
@@ -453,6 +513,9 @@ class AKSPreviewAgentPoolAddDecoratorStandaloneModeTestCase(AKSPreviewAgentPoolA
 
     def test_set_up_custom_ca_trust(self):
         self.common_set_up_custom_ca_trust()
+
+    def test_set_up_artifact_streaming(self):
+        self.common_set_up_artifact_streaming()
 
     def test_construct_agentpool_profile_preview(self):
         import inspect
@@ -546,6 +609,9 @@ class AKSPreviewAgentPoolAddDecoratorManagedClusterModeTestCase(AKSPreviewAgentP
 
     def test_set_up_custom_ca_trust(self):
         self.common_set_up_custom_ca_trust()
+
+    def test_set_up_artifact_streaming(self):
+        self.common_set_up_artifact_streaming()
 
     def test_construct_agentpool_profile_preview(self):
         import inspect
@@ -668,7 +734,7 @@ class AKSPreviewAgentPoolUpdateDecoratorCommonTestCase(unittest.TestCase):
         )
         # fail on passing the wrong agentpool object
         with self.assertRaises(CLIInternalError):
-            dec_1.update_vm_properties(None)
+            dec_1.update_custom_ca_trust(None)
         agentpool_1 = self.create_initialized_agentpool_instance(
             enable_custom_ca_trust=False,
         )
@@ -688,7 +754,7 @@ class AKSPreviewAgentPoolUpdateDecoratorCommonTestCase(unittest.TestCase):
         )
         # fail on passing the wrong agentpool object
         with self.assertRaises(CLIInternalError):
-            dec_2.update_vm_properties(None)
+            dec_2.update_custom_ca_trust(None)
         agentpool_2 = self.create_initialized_agentpool_instance(
             enable_custom_ca_trust=True,
         )
@@ -696,6 +762,51 @@ class AKSPreviewAgentPoolUpdateDecoratorCommonTestCase(unittest.TestCase):
         dec_agentpool_2 = dec_2.update_custom_ca_trust(agentpool_2)
         grond_truth_agentpool_2 = self.create_initialized_agentpool_instance(
             enable_custom_ca_trust=False,
+        )
+        self.assertEqual(dec_agentpool_2, grond_truth_agentpool_2)
+
+    def common_update_artifact_streaming(self):
+        dec_1 = AKSPreviewAgentPoolUpdateDecorator(
+            self.cmd,
+            self.client,
+            {"enable_artifact_streaming": None},
+            self.resource_type,
+            self.agentpool_decorator_mode,
+        )
+        # fail on passing the wrong agentpool object
+        with self.assertRaises(CLIInternalError):
+            dec_1.update_artifact_streaming(None)
+        agentpool_1 = self.create_initialized_agentpool_instance(
+            artifact_streaming_profile=self.models.AgentPoolArtifactStreamingProfile(
+                enabled=True
+            )
+        )
+        dec_1.context.attach_agentpool(agentpool_1)
+        dec_agentpool_1 = dec_1.update_artifact_streaming(agentpool_1)
+        grond_truth_agentpool_1 = self.create_initialized_agentpool_instance(
+            artifact_streaming_profile=self.models.AgentPoolArtifactStreamingProfile(
+                enabled=True
+            )
+        )
+        self.assertEqual(dec_agentpool_1, grond_truth_agentpool_1)
+
+        dec_2 = AKSPreviewAgentPoolUpdateDecorator(
+            self.cmd,
+            self.client,
+            {"enable_artifact_streaming": True},
+            self.resource_type,
+            self.agentpool_decorator_mode,
+        )
+        # fail on passing the wrong agentpool object
+        with self.assertRaises(CLIInternalError):
+            dec_2.update_artifact_streaming(None)
+        agentpool_2 = self.create_initialized_agentpool_instance()
+        dec_2.context.attach_agentpool(agentpool_2)
+        dec_agentpool_2 = dec_2.update_artifact_streaming(agentpool_2)
+        grond_truth_agentpool_2 = self.create_initialized_agentpool_instance(
+            artifact_streaming_profile=self.models.AgentPoolArtifactStreamingProfile(
+                enabled=True
+            )
         )
         self.assertEqual(dec_agentpool_2, grond_truth_agentpool_2)
 
@@ -713,6 +824,9 @@ class AKSPreviewAgentPoolUpdateDecoratorStandaloneModeTestCase(AKSPreviewAgentPo
 
     def test_update_custom_ca_trust(self):
         self.common_update_custom_ca_trust()
+
+    def test_update_artifact_streaming(self):
+        self.common_update_artifact_streaming()
 
     def test_update_agentpool_profile_preview(self):
         import inspect
@@ -776,6 +890,9 @@ class AKSPreviewAgentPoolUpdateDecoratorManagedClusterModeTestCase(AKSPreviewAge
 
     def test_update_custom_ca_trust(self):
         self.common_update_custom_ca_trust()
+
+    def test_update_artifact_streaming(self):
+        self.common_update_artifact_streaming()
 
     def test_update_agentpool_profile_preview(self):
         import inspect